Prismarine: Improve TTY support and remove `EvalCommand`.
We should further refactor the `CommandManager` to not dynamically
load commands as this makes everything way more complicated than it
needs to be it also currently (somewhat) breaks `cjs` support.

The `EvalCommand` was removed as it was almost certainly going to be
if not already a security risk. Something similar but in a virtual
environment could be added in the future for both debugging and map-
making purposes.