Prismarine: Remove plugin support.
"the sword that kills is the sword that gives life", sometimes you
have to burn some things down for something better to take place;
this is one of those times. As a part of the large refactor I will be
reworking plugins completely; but that depends on how we choose to
impl packets, handlers etc. So for now so long plugin support. 404ccbe84216d8ccec23254b6c0b8c25a0f69727