Prismarine+Server+Client: Add `logger` package.
This removes the logger from the `prismarine` package by breaking it
out into its own package. This is to allow for more flexibility and
way better testing.

We also export the custom transport so that it can be used in other
projects if anybody sees fit. :^) 941fb74a7818afd21e87804f62e004cf8465e0c4