Meta+Nightly: Build project before trying to package it.
Looks like bun is failing to find some of the packages in the monorepo
when trying to package the project. That is probably because it's
working-directory is in the `@JSPrismarine/server` package, so this
might fix it; but who knows! 505c720a00f17b0232e4ba143b5fdb9ef736452e